requirements.txt
cerebras_cloud_sdk
python-dotenv
fal-client
pyyaml
pynput
elevenlabs
pygame
pyaudio
gpiozero